chore(turbopack): Add `comment_width = 100` to rustfmt.toml (#68541)
The default `comment_width` is `80`, which is a mismatch with the
`max_width` of `100`.

Previous discussion:
https://vercel.slack.com/archives/C03EWR7LGEN/p1716332166435329

Ran `cargo fmt` after making this change. It seems to mostly just fix
bullet points. Other places it likely can't differentiate between
wrapped text and intentional newlines.

I don't plan on going through and manually re-wrapping of all of the
comments, but over time this should get fixed as people change comments.