refactor(turbopack): Remove cast to `EcmascriptModuleAsset` (#68427)
### What?

Now we have tree shaking implementation but it uses
`EcmascriptModulePartAsset` as the module type. So casting to
`EcmascriptModuleAsset` can break the tree shaking.

This PR also removes `special exports` of turbopack, which is
cumbersome. The previous analysis logic for those special exports may
depend on the value being `EcmascriptModuleAsset`, which is resolved in
this PR.

### Why?

Clean code & for future woork.


### How?

Closes PACK-3175